+++ scipy/optimize/src/nnls.h
@@ -2,10 +2,11 @@
 #define NNLS_H
 
 #include <math.h>
-#include "npy_cblas.h"
+#include <stdint.h>
+//#include "npy_cblas.h"
 #include "blaslapack_declarations.h"
 
-void __nnls(const CBLAS_INT m, const CBLAS_INT n, double* restrict a, double* restrict b, double* restrict x, double* restrict w, double* restrict zz, CBLAS_INT* restrict indices, const Py_ssize_t maxiter, double* rnorm, int* info);
+void __nnls(const CBLAS_INT m, const CBLAS_INT n, double* restrict a, double* restrict b, double* restrict x, double* restrict w, double* restrict zz, CBLAS_INT* restrict indices, const int64_t maxiter, double* rnorm, int64_t* info);
 
 
 #endif
